🔧 [Settings] 기본 구조 세팅 및 패키지 분리
